After a 'rustup update', there is a new round of clippy improvements.
1) Can removed a rounding error by using a mul_add().
2) Idomatic use of format!() macro.
3) Removed needless conversion of f64.
4) Default did not follow the definition exactly.

examples/globe/drag_and_scale minor tinkering.

Adjusted scale limits.
Adjusted color scheme.

+++ examples/globe/rotating_50m/src/lib.rs
@@ -119,7 +119,7 @@ impl Renderer {
             feature_from_name(&topology, "countries").expect("Did not extract geometry");
 
         let mut ob = Orthographic::builder();
-        ob.scale_set(width as f64 / 1.3_f64 / std::f64::consts::PI)
+        ob.scale_set(width / 1.3_f64 / std::f64::consts::PI)
             .translate_set(&Coord {
                 x: width / 2_f64,
                 y: height / 2_f64,
