+++ examples/projections/src/azimuthal_equal_area.rs
@@ -36,7 +36,7 @@ pub async fn draw_azimuthal_equal_area(land: &Geometry<f64>) -> Result<(), JsVal
     let pb = PathBuilder::new(context);
 
     let azimuthal_equal_area = AzimuthalEqualArea::builder()
-        .scale_set(width as f64 / 3_f64)
+        .scale_set(width / 3_f64)
         .translate_set(&Coord {
             x: width / 2_f64,
             y: height / 2_f64,
