+++ examples/projections/src/mercator.rs
@@ -37,7 +37,7 @@ pub async fn draw_mercator(land: &Geometry<f64>) -> Result<(), JsValue> {
     let pb = PathBuilder::new(context);
 
     let mut mercator = Mercator::builder();
-    let mercator = mercator.scale_set(width as f64 / 1.3_f64 / std::f64::consts::PI);
+    let mercator = mercator.scale_set(width / 1.3_f64 / std::f64::consts::PI);
     mercator.translate_set(&Coord {
         x: width / 2_f64,
         y: height / 2_f64,
