+++ examples/projections/src/mercator_transverse.rs
@@ -39,7 +39,7 @@ pub async fn draw_mercator_transverse(land: &Geometry<f64>) -> Result<(), JsValu
     let mut mercator = MercatorTransverse::builder();
     let mercator = mercator
         .base
-        .scale_set(width as f64 / 1.3_f64 / std::f64::consts::PI);
+        .scale_set(width / 1.3_f64 / std::f64::consts::PI);
     mercator.translate_set(&Coord {
         x: width / 2_f64,
         y: height / 2_f64,
