+++ examples/projections/src/orthographic.rs
@@ -35,7 +35,7 @@ pub async fn draw_orthographic(land: &Geometry<f64>) -> Result<(), JsValue> {
     let pb = PathBuilder::new(context);
 
     let ortho = Orthographic::builder()
-        .scale_set(width as f64 / 1.3_f64 / std::f64::consts::PI)
+        .scale_set(width / 1.3_f64 / std::f64::consts::PI)
         .translate_set(&Coord {
             x: width / 2_f64,
             y: height / 2_f64,
